+++ src/components/Home/ActivityTicker.css
@@ -10,10 +10,6 @@
   overflow: hidden;
 }
 
-.ActivityTicker__EmptyContainer {
-  margin-top: 32px;
-}
-
 .ActivityTicker__TitleContainer {
   position: relative;
   background-color: transparent;
@@ -37,36 +33,6 @@
   width: 100%;
   height: 27px;
   background: linear-gradient(180deg, #fff 82.84%, rgba(255, 255, 255, 0) 100%);
-}
-
-.ActivityTicker__LoadingContainer {
-  position: relative;
-  top: 100px;
-}
-
-.ActivityTicker__List {
-  display: flex;
-  flex-direction: column;
-  overflow: scroll;
-  overflow-x: hidden;
-  overscroll-behavior-y: contain;
-  padding-top: 8px;
-}
-
-.ActivityTicker__List::-webkit-scrollbar {
-  display: none;
-}
-
-.ActivityTicker__ListItem {
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  padding: 10px;
-  border-radius: 8px;
-}
-
-.ActivityTicker__ListItem:hover {
-  background-color: var(--black-200);
 }
 
 .ActivityTicker__ListItemMarkdown {
